+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 * BUGFIX keep shortcodes inside of widgets, let theme function handle to do them or not
+* Enable shortcodes inside of testimonials content
 
 ## 2.13.2
 * Add StillMaintained.com notice